refactor: 레디스 사용해보기
